requirements.txt
Django==4.2.7
gunicorn==21.2.0
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
googlemaps==4.10.0
webdriver-manager==4.0.1
urllib3==2.1.0
python-dotenv==1.0.0
fake-useragent==1.4.0
extruct==0.13.0
phonenumbers==8.13.26
validators==0.22.0
python-whois==0.8.0
aiohttp==3.9.1
tldextract==5.1.0

